made some changes and added expense

+++ StudentAllowanceTracker.Application/Commands/Allowances/DeleteAllowanceCommandHandler.cs
@@ -6,24 +6,27 @@ using System.Text;
 using System.Threading.Tasks;
 using StudentAllowanceTracker.Domain.Interfaces.Repositories;
 using StudentAllowanceTracker.Domain.Entities;
+using StudentAllowanceTracker.Shared.Enums;
+using StudentAllowanceTracker.Shared.Responses;
+using StudentAllowanceTracker.Application.DTOs;
+
 namespace StudentAllowanceTracker.Application.Commands.Allowances
 {
-    public class DeleteAllowanceCommandHandler: IRequestHandler<DeleteAllowanceCommand>
+    public class DeleteAllowanceCommandHandler: IRequestHandler<DeleteAllowanceCommand, Result<object> >
     {
         private readonly IBaseRepository<Allowance> _allowanceRepo;
         public DeleteAllowanceCommandHandler(IBaseRepository<Allowance> allowanceRepo)
         {
             _allowanceRepo = allowanceRepo;
         }
-
-        public async Task Handle(DeleteAllowanceCommand request, CancellationToken cancellationToken)
+        public async Task<Result<object>> Handle(DeleteAllowanceCommand request, CancellationToken cancellationToken)
         {
-            var allowance = await _allowanceRepo.GetByIdAsync(request.AllowanceID)
-                          ?? throw new Exception("Allowance not found.");
+            var allowance = await _allowanceRepo.GetByIdAsync(request.AllowanceID);
+            if (allowance == null)
+                return Result<object>.Fail(ResultStatus.NotFound, "Allowance not found.");
 
             await _allowanceRepo.DeleteAsync(request.AllowanceID);
-
-
+            return Result<object>.Ok();
         }
     }
 }
